+++ src/test/java/de/uni_jena/cs/fusion/abecto/processor/AbstractValueComparisonProcessorTest.java
@@ -19,6 +19,7 @@ import static de.uni_jena.cs.fusion.abecto.TestUtil.aspect;
 import static de.uni_jena.cs.fusion.abecto.TestUtil.containsDeviation;
 import static de.uni_jena.cs.fusion.abecto.TestUtil.containsIssue;
 import static de.uni_jena.cs.fusion.abecto.TestUtil.containsValuesOmission;
+import static de.uni_jena.cs.fusion.abecto.TestUtil.containsualityMeasurement;
 import static de.uni_jena.cs.fusion.abecto.TestUtil.dataset;
 import static de.uni_jena.cs.fusion.abecto.TestUtil.property;
 import static de.uni_jena.cs.fusion.abecto.TestUtil.resource;
@@ -27,6 +28,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import java.math.BigDecimal;
+import java.math.RoundingMode;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
@@ -42,6 +45,7 @@ import org.apache.jena.vocabulary.RDF;
 
 import de.uni_jena.cs.fusion.abecto.Aspect;
 import de.uni_jena.cs.fusion.abecto.vocabulary.AV;
+import de.uni_jena.cs.fusion.abecto.vocabulary.OM;
 
 public abstract class AbstractValueComparisonProcessorTest {
 
@@ -50,7 +54,7 @@ public abstract class AbstractValueComparisonProcessorTest {
 	Aspect aspect1 = new Aspect(aspect(1), "key").setPattern(dataset(1), pattern).setPattern(dataset(2), pattern);
 	Model mappingModel = ModelFactory.createDefaultModel();
 	{
-		addMapping(subject(1), subject(2));
+		addMapping(subject(1), subject(2), subject(3));
 	}
 
 	public void addMapping(Resource... resources) {
@@ -90,6 +94,9 @@ public abstract class AbstractValueComparisonProcessorTest {
 		assertEquals(1, outputMetaModels[0].listStatements(null, RDF.type, AV.Issue).toList().size());
 		assertEquals(0, outputMetaModels[1].listStatements(null, RDF.type, AV.Issue).toList().size());
 
+		assertMeasurements(dataset(1), dataset(2), aspect(1), "value", outputMetaModels[0], 1, 1, 0, 0);
+		assertMeasurements(dataset(2), dataset(1), aspect(1), "value", outputMetaModels[1], 1, 1, 0, 0);
+
 		// second direction
 		model1 = ModelFactory.createDefaultModel().add(subject(1), property(1), expectedValue);
 		model2 = ModelFactory.createDefaultModel().add(subject(2), property(1), unexpectedValue);
@@ -105,6 +112,9 @@ public abstract class AbstractValueComparisonProcessorTest {
 				outputMetaModels[1]));
 		assertEquals(0, outputMetaModels[0].listStatements(null, RDF.type, AV.Issue).toList().size());
 		assertEquals(1, outputMetaModels[1].listStatements(null, RDF.type, AV.Issue).toList().size());
+
+		assertMeasurements(dataset(1), dataset(2), aspect(1), "value", outputMetaModels[0], 1, 1, 0, 0);
+		assertMeasurements(dataset(2), dataset(1), aspect(1), "value", outputMetaModels[1], 1, 1, 0, 0);
 	}
 
 	void assertDeviation(RDFNode value1, RDFNode value2) throws Exception {
@@ -113,6 +123,12 @@ public abstract class AbstractValueComparisonProcessorTest {
 	}
 
 	void assertDeviation(Collection<RDFNode> values1, Collection<RDFNode> values2,
+			Collection<RDFNode> notDeviatingValues1, Collection<RDFNode> notDeviatingValues2) throws Exception {
+		assertDeviationOneDirection(values1, values2, notDeviatingValues1, notDeviatingValues2);
+		assertDeviationOneDirection(values2, values1, notDeviatingValues2, notDeviatingValues1);
+	}
+
+	void assertDeviationOneDirection(Collection<RDFNode> values1, Collection<RDFNode> values2,
 			Collection<RDFNode> notDeviatingValues1, Collection<RDFNode> notDeviatingValues2) throws Exception {
 		int expectedDeviationCount = (values1.size() - notDeviatingValues1.size())
 				* (values2.size() - notDeviatingValues2.size());
@@ -150,42 +166,18 @@ public abstract class AbstractValueComparisonProcessorTest {
 		assertEquals(expectedDeviationCount,
 				outputMetaModels[1].listStatements(null, RDF.type, AV.Deviation).toList().size());
 
-		// second direction
-		model1 = ModelFactory.createDefaultModel();
-		model2 = ModelFactory.createDefaultModel();
-		for (RDFNode value2 : values2) {
-			model1.add(subject(1), property(1), value2);
-		}
-		for (RDFNode value1 : values1) {
-			model2.add(subject(2), property(1), value1);
-		}
-		model1.add(subject(1), property(2), resource("alwaysPresent"));
-		model2.add(subject(2), property(2), resource("alwaysPresent"));
-		outputMetaModels = compare(model1, model2);
-		assertFalse(outputMetaModels[0].contains(null, RDF.type, AV.Issue));
-		assertFalse(outputMetaModels[1].contains(null, RDF.type, AV.Issue));
-		assertFalse(outputMetaModels[0].contains(null, RDF.type, AV.ValueOmission));
-		assertFalse(outputMetaModels[1].contains(null, RDF.type, AV.ValueOmission));
-
-		for (RDFNode value1 : values1) {
-			if (!notDeviatingValues1.contains(value1)) {
-				for (RDFNode value2 : values2) {
-					if (!notDeviatingValues2.contains(value2)) {
-						assertTrue(containsDeviation(subject(1), "value", value2, dataset(2), subject(2), value1,
-								aspect(1), outputMetaModels[0]));
-						assertTrue(containsDeviation(subject(2), "value", value1, dataset(1), subject(1), value2,
-								aspect(1), outputMetaModels[1]));
-					}
-				}
-			}
-		}
-		assertEquals(expectedDeviationCount,
-				outputMetaModels[0].listStatements(null, RDF.type, AV.Deviation).toList().size());
-		assertEquals(expectedDeviationCount,
-				outputMetaModels[1].listStatements(null, RDF.type, AV.Deviation).toList().size());
+		assertMeasurements(dataset(1), dataset(2), aspect(1), "value", outputMetaModels[0], values1.size(),
+				values2.size(), notDeviatingValues2.size(), notDeviatingValues1.size());
+		assertMeasurements(dataset(2), dataset(1), aspect(1), "value", outputMetaModels[1], values2.size(),
+				values1.size(), notDeviatingValues1.size(), notDeviatingValues2.size());
 	}
 
 	void assertSame(RDFNode value1, RDFNode value2) throws Exception {
+		assertSameOneDirection(value2, value1);
+		assertSameOneDirection(value1, value2);
+	}
+
+	void assertSameOneDirection(RDFNode value1, RDFNode value2) throws Exception {
 		// first direction
 		Model model1 = ModelFactory.createDefaultModel().add(subject(1), property(1), value1);
 		Model model2 = ModelFactory.createDefaultModel().add(subject(2), property(1), value2);
@@ -198,24 +190,19 @@ public abstract class AbstractValueComparisonProcessorTest {
 		assertFalse(outputMetaModels[1].contains(null, RDF.type, AV.Deviation));
 		assertFalse(outputMetaModels[0].contains(null, RDF.type, AV.ValueOmission));
 		assertFalse(outputMetaModels[1].contains(null, RDF.type, AV.ValueOmission));
-
-		// second direction
-		model1 = ModelFactory.createDefaultModel().add(subject(1), property(1), value2);
-		model2 = ModelFactory.createDefaultModel().add(subject(2), property(1), value1);
-		model1.add(subject(1), property(2), resource("alwaysPresent"));
-		model2.add(subject(2), property(2), resource("alwaysPresent"));
-		outputMetaModels = compare(model1, model2);
-		assertFalse(outputMetaModels[0].contains(null, RDF.type, AV.Issue));
-		assertFalse(outputMetaModels[1].contains(null, RDF.type, AV.Issue));
-		assertFalse(outputMetaModels[0].contains(null, RDF.type, AV.Deviation));
-		assertFalse(outputMetaModels[1].contains(null, RDF.type, AV.Deviation));
-		assertFalse(outputMetaModels[0].contains(null, RDF.type, AV.ValueOmission));
-		assertFalse(outputMetaModels[1].contains(null, RDF.type, AV.ValueOmission));
+		// check measurement
+		assertMeasurements(dataset(1), dataset(2), aspect(1), "value", outputMetaModels[0], 1, 1, 1, 1);
+		assertMeasurements(dataset(2), dataset(1), aspect(1), "value", outputMetaModels[1], 1, 1, 1, 1);
 	}
 
 	void assertMissing(Collection<RDFNode> values1, Collection<RDFNode> values2, Collection<RDFNode> missingValues1,
 			Collection<RDFNode> missingValues2) throws Exception {
-		// first direction
+		assertMissingOneDirection(values1, values2, missingValues1, missingValues2);
+		assertMissingOneDirection(values2, values1, missingValues2, missingValues1);
+	}
+
+	void assertMissingOneDirection(Collection<RDFNode> values1, Collection<RDFNode> values2,
+			Collection<RDFNode> missingValues1, Collection<RDFNode> missingValues2) throws Exception {
 		Model model1 = ModelFactory.createDefaultModel();
 		Model model2 = ModelFactory.createDefaultModel();
 		for (RDFNode value1 : values1) {
@@ -243,34 +230,43 @@ public abstract class AbstractValueComparisonProcessorTest {
 			assertTrue(containsValuesOmission(subject(2), "value", dataset(1), subject(1), missingValue2, aspect(1),
 					outputMetaModels[1]));
 		}
+		// check measurement
+		assertMeasurements(dataset(1), dataset(2), aspect(1), "value", outputMetaModels[0], values1.size(),
+				values2.size(), values2.size() - missingValues1.size(), values1.size() - missingValues2.size());
+		assertMeasurements(dataset(2), dataset(1), aspect(1), "value", outputMetaModels[1], values2.size(),
+				values1.size(), values1.size() - missingValues2.size(), values2.size() - missingValues1.size());
+	}
 
-		// second direction
-		model1 = ModelFactory.createDefaultModel();
-		model2 = ModelFactory.createDefaultModel();
-		for (RDFNode value2 : values2) {
-			model1.add(subject(1), property(1), value2);
+	void assertMeasurements(Resource dataset1, Resource dataset2, Resource aspect, String variable,
+			Model outputMetaModel, int count1, int count2, int absoluteCoverage1, int absoluteCoverage2)
+			throws Exception {
+		BigDecimal relativeCoverage = BigDecimal.ZERO;
+		if (count2 != 0) {
+			relativeCoverage = BigDecimal.valueOf(absoluteCoverage1).divide(BigDecimal.valueOf(count2), 2,
+					RoundingMode.HALF_UP);
 		}
-		for (RDFNode value1 : values1) {
-			model2.add(subject(2), property(1), value1);
+		BigDecimal totalPairwiseOverlap = BigDecimal.valueOf(absoluteCoverage1 + absoluteCoverage2)
+				.divide(BigDecimal.valueOf(2), 0, RoundingMode.CEILING);
+		BigDecimal populationSize = null;
+		if (!totalPairwiseOverlap.equals(BigDecimal.ZERO)) {
+			populationSize = BigDecimal.valueOf(count1 * count2).divide(totalPairwiseOverlap, 2, RoundingMode.HALF_UP);
 		}
-		model1.add(subject(1), property(2), resource("alwaysPresent"));
-		model2.add(subject(2), property(2), resource("alwaysPresent"));
-		outputMetaModels = compare(model1, model2);
-		assertFalse(outputMetaModels[0].contains(null, RDF.type, AV.Issue));
-		assertFalse(outputMetaModels[1].contains(null, RDF.type, AV.Issue));
-		assertFalse(outputMetaModels[0].contains(null, RDF.type, AV.Deviation));
-		assertFalse(outputMetaModels[1].contains(null, RDF.type, AV.Deviation));
-		assertEquals(missingValues2.size(),
-				outputMetaModels[0].listStatements(null, RDF.type, AV.ValueOmission).toList().size());
-		assertEquals(missingValues1.size(),
-				outputMetaModels[1].listStatements(null, RDF.type, AV.ValueOmission).toList().size());
-		for (RDFNode missingValue2 : missingValues2) {
-			assertTrue(containsValuesOmission(subject(1), "value", dataset(2), subject(2), missingValue2, aspect(1),
-					outputMetaModels[0]));
+		BigDecimal completeness = null;
+		if (populationSize != null) {
+			completeness = BigDecimal.valueOf(count1).divide(populationSize, 2, RoundingMode.HALF_UP);
 		}
-		for (RDFNode missingValue1 : missingValues1) {
-			assertTrue(containsValuesOmission(subject(2), "value", dataset(1), subject(1), missingValue1, aspect(1),
-					outputMetaModels[1]));
+		assertTrue(containsualityMeasurement(AV.count, count1, OM.one, dataset1, variable, Collections.emptyList(),
+				aspect(1), outputMetaModel));
+		assertTrue(containsualityMeasurement(AV.absoluteCoverage, absoluteCoverage1, OM.one, dataset1, variable,
+				Collections.singleton(dataset2), aspect, outputMetaModel));
+		assertTrue(containsualityMeasurement(AV.relativeCoverage, relativeCoverage, OM.one, dataset1, variable,
+				Collections.singleton(dataset2), aspect, outputMetaModel));
+		if (completeness != null) {
+			assertTrue(containsualityMeasurement(AV.marCompletenessThomas08, completeness, OM.one, dataset1, variable,
+					Collections.singleton(dataset2), aspect, outputMetaModel));
+		} else {
+			assertFalse(containsualityMeasurement(AV.marCompletenessThomas08, null, OM.one, dataset1, variable,
+					Collections.singleton(dataset2), aspect, outputMetaModel));
 		}
 	}
 }